backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.4.26
chardet==4.0.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
fastmcp==2.2.6
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-oauth2-tool==0.0.3
googleapis-common-protos==1.70.0
gql==3.5.2
graphql-core==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==2.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.58
langchain-mcp-adapters==0.0.9
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.41
linear-py==0.0.6
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
multidict==6.4.3
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.4.8
openapi-pydantic==0.5.1
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.25.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
tenacity==9.1.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.2
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
